Populate multiple form fields for address
I'm trying to use this to autocomplete an address form when a person is filling out their address. I got it to work in <code>geocode_lookup()</code> but not in <code>autocomplete_init()</code>. So basically if you select one of the options from the drop down with your mouse it uses <code>autocomplete_init()</code>. I can't figure out how to pass the correct parameters to <code>update_ui()</code> inside of <code>autocomplete_init()</code> but I've managed to do it in <code>geocode_lookup()</code>.

I'm not to familiar with navigating my way through jQuery so it's probably something fairly simple that I'm missing.